+++ src/detect-dns-response.c
@@ -25,6 +25,7 @@
 #include "detect-parse.h"
 #include "detect-engine.h"
 #include "detect-engine-buffer.h"
+#include "detect-engine-mpm.h"
 #include "detect-engine-prefilter.h"
 #include "detect-engine-content-inspection.h"
 #include "detect-dns-response.h"
